ktesting: reorganize tests
Tests get moved into files matching the implementation.
Test coverage is at 89.6% with or without this change, so
no test got lost.
